+++ ematoblender/scripts/ema_io/ema_staticserver/mocap_file_parser_old.py
@@ -11,173 +11,9 @@ import xml.etree.ElementTree as ET
 import math
 import struct
 import os
-from ematoblender.scripts.ema_io.rtc3d_parser import DataFrame, Component6D, CoilBuilder6D
+from ematoblender.scripts.ema_io.rtc3d_parser import DataFrame, Component, Coil
 
-xml_skeleton_location = os.path.abspath(os.path.normpath(os.path.dirname(__file__)+'./parameter_skeleton.xml'))
-
-
-class Mapping():
-    """Control the reordering between the mocap file and the WAVE order."""
-    def __init__(self, names_as_strings=None, mapping_as_list=None, angle_type='QUATERNION'):
-        self.quat_inds = []
-        self.euler_xyz = []
-
-        if names_as_strings is not None:
-            mapping_as_list = self.order_measurements_as_wave(names_as_strings, nchannels=6)
-        if angle_type == 'QUATERNION':  # quaternions are OK
-            if mapping_as_list is not None:
-                self.q0ind, self.qxind, self.qyind, self.qzind, self.xind, self.yind, self.zind = mapping_as_list
-            else:
-                self.q0ind, self.qxind, self.qyind, self.qzind, self.xind, self.yind, self.zind = [i for i in range(7)]
-                self.otherind = 8
-
-            self.quat_inds = [self.q0ind, self.qxind, self.qyind, self.qzind]
-
-        if angle_type == 'ELEVATION':
-            self.quat_inds = [None]
-            if mapping_as_list is not None:
-                a, pi, theta, d, self.xind, self.yind, self.zind = mapping_as_list
-            self.euler_xyz = [theta, 0,  pi]
-
-
-
-    @staticmethod
-    def make_float_or_zero(listofvals):
-        """Converts all items in the list to floats, or to 0.0 if not possible."""
-        floatvals = []
-        for v in listofvals:
-            try:
-                v = float(v)
-            except ValueError:
-                v = 0.0
-            finally:
-                floatvals.append(v)
-        return floatvals
-
-    def convert_to_quat(self, measurementlist):
-        """Convert the angle information in the mapping into Quaternions
-        Return q0, qx, qy, qz.
-        """
-        print('converting {} to quats'.format(measurementlist))
-        #TODO: Handle the cases where quaternions are not given
-        if all(x is not None for x in self.quat_inds) and all([measurementlist[x] is not None for x in self.quat_inds]):
-            print(self.make_float_or_zero([measurementlist[i] for i in self.quat_inds]))
-            return self.make_float_or_zero([measurementlist[i] for i in self.quat_inds])
-
-        elif all([measurementlist[x] is not None for x in self.euler_xyz]):
-            x, y, z = [measurementlist[x] is not None for x in self.euler_xyz]
-            # TODO: This presumes the angle application order XYZ # TODO: Check source
-            # TODO: Source is http://www.euclideanspace.com/maths/geometry/rotations/conversions/eulerToQuaternion/
-
-            c1, c2, c3 = [math.cos(math.radians(i)/2) for i in [z, y, x]]  # z is yaw, y is pitch, x is roll
-            s1, s2, s3 = [math.sin(math.radians(i)/2) for i in [z, y, x]]  # z is yaw, y is pitch, x is roll
-
-            # convert euler rotation angles to quaternion
-            q0 = c1*c2*c3 - s1*s2*s3
-            qx = s1*s2*c3 + c1*c2*s3
-            qy = s1*c2*c3 + c1*s2*s3
-            qz = c1*s2*c3 - s1*c2*s3
-            return self.make_float_or_zero((q0, qx, qy, qz))
-        else:
-            raise ValueError('Cannot convert None values to quaternions')
-
-    def give_in_wave_order(self, measurementlist):
-        """Given some measurement list that conforms to the mapping, return the measurements re-ordered."""
-        measurementlist += [0 for j in range(7)]  # pad in case too short
-        return [measurementlist[i] if i is not None else 0 for i in
-                [self.q0ind, self.qxind, self.qyind, self.qzind, self.xind, self.yind, self.zind]]
-
-
-    @staticmethod
-    def channel_label_to_attrs(label):
-        """Semantically parse label names."""
-        label = label.lower()
-
-        # rotation vs location
-        if 'rot' in label or 'q' in label: # rotation label
-            location = False; rotation = True
-        elif 'loc' in label or 'pos' in label or 'mm' in label: # position label
-            location = True
-            rotation = False
-        else:
-            location = False
-            rotation = False
-        zero = False
-        if '0' in label:
-            zero = True
-
-        # axis
-        if label.startswith('x') or label.endswith('x'):
-            x, y, z = True, False, False
-        elif label.startswith('y') or label.endswith('y'):
-            x, y, z = False, True, False
-        elif label.startswith('z') or label.endswith('z'):
-            x, y, z = False, False, True
-        else:
-            x, y, z = False, False, False
-
-        # Euler vs Quaternion
-        if label.startswith('q') or label.endswith('q'):
-            quat = True
-        else:
-            quat = False # Euler angles presumed
-        return location, rotation, x, y, z, quat, zero
-
-    @staticmethod
-    def order_measurements_as_wave(labels, nchannels=3):
-        """
-        Takes a list of labels that describe the output from the datafile.
-        That is, a list based on the wave protocol, with for each position, the index of the element in the input it should access.
-        Label order is: q0, qx, qy, qz, x, y, z
-        :return: List of indices that should be applied to the labels to re-order them for the WAVE protocol.
-        None for labels that aren't included in the final object.
-        """
-
-        origlen = len(labels)
-        attributes = [(i, Mapping.channel_label_to_attrs(l)) for i, l in enumerate(labels)]
-
-        # get lists for which channels are (guessed to be) location/rotation, x/y/z etc.
-        locs = [i for i, attrs in attributes if attrs[0]]
-        rots = [i for i, attrs in attributes if attrs[1]]
-
-        x = [i for i, attrs in attributes if attrs[2]]
-        y = [i for i, attrs in attributes if attrs[3]]
-        z = [i for i, attrs in attributes if attrs[4]]
-        q = [i for i, attrs in attributes if attrs[5]]
-        zero = [i for i, attrs in attributes if attrs[6]]
-
-        if nchannels >= 6:  # rotation available
-            # x, y, z, by intersection of loc/rots and xyz
-            labelindices = [[j for j in letter if j in orientation]
-                            for orientation in [rots, locs] for letter in [x,y,z] ] # first orientation is impossible as placeholder for q0
-            #print("returning the correct order as:", [labels[li[0]] for li in labelindices])
-            # make indices of length 1 else None, increment by one to allow place for q0
-
-            # return the reordered label indices
-            changed_indices = [li[0] if len(li)> 0  else None for li in labelindices]
-            #print("ci", changed_indices)
-
-            q0_refs = [i for i in q if i in zero]
-            q_index = min(q0_refs) if len(q0_refs)>0 else -1
-            #print("q index", q_index)
-            changed_indices = [q_index] + changed_indices  # q0 has a Q but no X Y or Z in label
-            #print("qi", changed_indices)
-
-            #print("original labels", labels)
-            #print("changed ind", changed_indices)
-
-        else:  # only location available for all, use 3d order
-
-            labelindices = [[j for j in letter if j in orientation] for orientation in [locs] for letter in [x,y,z] ]
-            #print("returning the correct order as:", [labels[li[0]]if len(li)>0 else None for li in labelindices])
-            changed_indices = [li[0] if len(li)>0 else None for li in labelindices[1:]]
-
-        # ensure all labels receive a new index
-        filling_none = [None for i in range(origlen - len(changed_indices))]
-        return changed_indices + filling_none   # return the reordered label indices
-
-
-
+xml_skeleton_location = './scripts/ema_io/ema_staticserver/parameter_skeleton.xml'
 
 
 class MakeMocapParser():
@@ -191,122 +27,75 @@ class MakeMocapParser():
             return TSVParser(filename)
         if extension == 'pos':
             return POSParser(filename)
-        if extension == 'json':
-            pass
-            #return JSONParser(filename)
-        raise ValueError("Bad mocap file type: {}".format(extension))
+        assert 0, "Bad mocap file type " + extension
     factory = staticmethod(factory)
 
 
 class MocapParent(object):
-    """Parent class to Mocap objects.
-    All Mocap file parser objects have the attributes:
-
-    FILE ATTRIBUTES
-    - file (the file object with the data)
-    - file_name (filename)
-    - format (extension)
-    - wave_name, video_name (file names of associated audio/video files)
-    - xml_tree (XML root object for parameter XML string (filled meaningfully later))
-
-    FRAME ATTRIBUTES
-    - motion_lines_read (position in reading by line)
-    - pre_motion_position (position where reading starts, bytes from start, after this can use readline() to get frames)
-    - max_num_frames (for static file, the number of frames that can  be read before looping)
-    - latest_timestamp (the most recently read timestamp, in microseconds)
-
-    SAMPLING RATE ATTRIBUTES
-    - sampling_rate (frames per second sampled)
-    - frame_time (number of microseconds per frame)
-    - frame_times (list of time difference between frames in seconds)
-
-    COIL ATTRIBUTES
-    - n_coils (number of EMA coils represented in file)
-    - marker_names (coil names as listed in file header)
-    - marker_channels (names of the measurements taken for each marker/coil in order (eg [xrot,  yrot, zrot, x, y, z]))
-        (note that this is presumed the same for every coil)
-
-    CONVERSION TO STANDARD FORMAT
-    - mappings (list indicating how the file's values should be reordered to emulate WAVE ordering )
-
-    ACTUAL LOCATION DATA
-    - component (component attribute as defined in rtc3d parser)
-
-    METHODS COMMON TO ALL MOCAP FORMATS:
-    - read_header (process the file header, populating attributes with the info)
-    - give_motion_frame (read and process one motion frame, returning a dataframe object with the info)
-    - update_xml_initial (update certain static xml attributes from the object's attributes)
-    - update_xml_stats (update XML to reflect the current state of the file reading)
-    - reset_motion_section (set file to position at start of the motion section)
-    - search_for_multimodal (scan the file's directory for likely audio/video candidates)
-    - order_measurements_as_wave (yield the mapping between the dimensions given per coil and the order
-        they should be in if they were streamed from the WAVE)
-
-    STATIC METHODS
-    - timestamp_to_microsecs
-    - make_float_or_zero
-    - channel_label_to_attrs (do NLP to gather whether name represents loc, rot, axis, format)
-
-    """
-
-    file_read_mode = 'r'
+    """Parent class to Mocap objects."""
 
     def __init__(self, filename):
         print('initialising a mocap parent')
 
         # open the file
         filename = filename if os.path.isabs(filename) else os.path.normpath(os.getcwd()+os.path.sep + filename)
-        self.file = open(filename, self.__class__.file_read_mode)
+        self.file = open(filename, 'r')
         self.file_name = self.file.name
-        self.format = os.path.splitext(filename)[-1]
-
-        self.xml_tree = ET.parse(xml_skeleton_location)
-        self.wave_name, self.video_name = self.search_for_multimodal()
+        self.format = filename.split('.')[-1]
 
         # capture the position within the file
         self.motion_lines_read = 0
-        self.pre_motion_position = NotImplemented
-        self.max_num_frames = NotImplemented
-        self.latest_timestamp = NotImplemented
+        self.pre_motion_position = 0  # positioned to readline() and give motion line
 
         # populated in subclasses from the header/pre-reading
-        self.sampling_rate = NotImplemented
-        self.frame_time = NotImplemented
-        self.frame_times = NotImplemented
+        self.sampling_rate = 0      # frames per second sampled
+        self.frame_time = 0         # number of microseconds per frame
+        self.frame_times = []       # list of time difference between frames in seconds
+        self.max_num_frames = 0     # for static file, the number of frames that can  be read before looping
 
         # perhaps these could be coil objects
-        self.n_coils = NotImplemented
-        self.marker_names = []
-        self.marker_channels = []
+        self.n_coils = 0            # integer, representing number of coils to be read
+        self.marker_names = []      # the labels in the static header for each coil, in order listed
+        self.marker_channels = []   # TODO: Tuple?
 
         # arbitrary large number,
         # later represents the number of dimensions measured to decide between 6d and 3d representations
         self.min_dimensions = 100
-        self.mappings = NotImplemented
+        self.mappings = None        # list indicating how the file's values should be reordered to emulate WAVE ordering
 
-        self.component = NotImplemented  # an initial component attribute
+        self.latest_timestamp = 0 # timestamp is microseconds
+        self.component = Component  # an initial component attribute
 
-        #  process the file header
-        self.read_header()
-        self.update_xml_initial()
+        self.xml_tree = ET.parse(xml_skeleton_location)  # basic standard XML root
+
+        self.wave_name, self.video_name = self.search_for_multimodal()
 
     @staticmethod
     def timestamp_to_microsecs(timestamp):
-        """
-        Get timestamp as string or int. Convert to ms.
-        """
-
+        """Get timestamp as string or int. Convert to ms."""
         # ts is a float, presume  it's in seconds
         if type(timestamp) == float or (type(timestamp) == str and timestamp.find('.')):
-            timestamp = math.floor(float(timestamp) * 1000000)  # seconds to microseconds
+            timestamp = math.floor(float(timestamp) * 1000000)  #seconds to microseconds
 
-        # ts is an integer or string of integer, presumably milliseconds. keep in same format
-        elif (type(timestamp) == str and timestamp.isnumeric()) or type(timestamp) == int:
+        # ts is an integer, presumably milliseconds. keep in same format
+        elif (type(timestamp) == str and timestamp.isnumeric()) or type(timestamp) == int: # presumably string of integer
             timestamp = int(timestamp)
         else:
             raise TypeError
         return timestamp
 
+    @staticmethod
+    def make_float_or_zero(listofvals):
+        """Converts all items in the list to floats, or to 0.0 if not possible."""
+        floatvals = []
+        for v in listofvals:
+            try:
+                v = float(v)
+            except ValueError:
+                v = 0.0
+            finally:
+                floatvals.append(v)
+        return floatvals
 
     def __str__(self):
         """String rep of the mocap file"""
@@ -395,6 +184,96 @@ class MocapParent(object):
 
         return wave_name, video_name
 
+    @staticmethod
+    def order_measurements_as_wave(labels, nchannels=3):
+        """
+        Takes a list of labels that describe the output from the datafile.
+        That is, a list based on the wave protocol, with for each position, the index of the element in the input it should access.
+        Label order is: q0, qx, qy, qz, x, y, z
+        :return: List of indices that should be applied to the labels to re-order them for the WAVE protocol.
+        None for labels that aren't included in the final object.
+        """
+
+        origlen = len(labels)
+        attributes = [(i, MocapParent.channel_label_to_attrs(l)) for i, l in enumerate(labels)]
+
+        # get lists for which channels are (guessed to be) location/rotation, x/y/z etc.
+        locs = [i for i, attrs in attributes if attrs[0]]
+        rots = [i for i, attrs in attributes if attrs[1]]
+
+        x = [i for i, attrs in attributes if attrs[2]]
+        y = [i for i, attrs in attributes if attrs[3]]
+        z = [i for i, attrs in attributes if attrs[4]]
+        q = [i for i, attrs in attributes if attrs[5]]
+        zero = [i for i, attrs in attributes if attrs[6]]
+
+        if nchannels >= 6:  # rotation available
+            # x, y, z, by intersection of loc/rots and xyz
+            labelindices = [[j for j in letter if j in orientation]
+                            for orientation in [rots, locs] for letter in [x,y,z] ] # first orientation is impossible as placeholder for q0
+            #print("returning the correct order as:", [labels[li[0]] for li in labelindices])
+            # make indices of length 1 else None, increment by one to allow place for q0
+
+            # return the reordered label indices
+            changed_indices = [li[0] if len(li)> 0  else None for li in labelindices]
+            #print("ci", changed_indices)
+
+            q0_refs = [i for i in q if i in zero]
+            q_index = min(q0_refs) if len(q0_refs)>0 else -1
+            #print("q index", q_index)
+            changed_indices = [q_index] + changed_indices  # q0 has a Q but no X Y or Z in label
+            #print("qi", changed_indices)
+
+            #print("original labels", labels)
+            #print("changed ind", changed_indices)
+
+        else:  # only location available for all, use 3d order
+
+            labelindices = [[j for j in letter if j in orientation] for orientation in [locs] for letter in [x,y,z] ]
+            #print("returning the correct order as:", [labels[li[0]]if len(li)>0 else None for li in labelindices])
+            changed_indices = [li[0] if len(li)>0 else None for li in labelindices[1:]]
+
+        # ensure all labels receive a new index
+        filling_none = [None for i in range(origlen - len(changed_indices))]
+        return changed_indices + filling_none   # return the reordered label indices
+
+
+    # unused in this application
+    @staticmethod
+    def channel_label_to_attrs(label):
+        """Semantically parse label names."""
+        label = label.lower()
+
+        # rotation vs location
+        if 'rot' in label or 'q' in label: # rotation label
+            location = False; rotation = True
+        elif 'loc' in label or 'pos' in label or 'mm' in label: # position label
+            location = True
+            rotation = False
+        else:
+            location = False
+            rotation = False
+        zero = False
+        if '0' in label:
+            zero = True
+
+        # axis
+        if label.startswith('x') or label.endswith('x'):
+            x, y, z = True, False, False
+        elif label.startswith('y') or label.endswith('y'):
+            x, y, z = False, True, False
+        elif label.startswith('z') or label.endswith('z'):
+            x, y, z = False, False, True
+        else:
+            x, y, z = False, False, False
+
+        # Euler vs Quaternion
+        if label.startswith('q') or label.endswith('q'):
+            quat = True
+        else:
+            quat = False # Euler angles presumed
+        return location, rotation, x, y, z, quat, zero
+
 
 class BVHParser(MocapParent):
     """
@@ -405,11 +284,14 @@ class BVHParser(MocapParent):
         """ Initialise the BVH Parser object. """
 
         super().__init__(filename)
+        self.read_header()
+        self.update_xml_initial()
 
         # get the mappings between the measurements and wave orderings
-        self.mappings = Mapping(names_as_strings=self.marker_channels[0])
+        self.mappings = [self.order_measurements_as_wave(channels, nchannels=self.min_dimensions)
+                         for channels in self.marker_channels]
 
-        self.component = Component6D(fileparser=self)  # builds coil objects within init
+        self.component = Component(fileparser=self)  # builds coil objects within init
 
     def __str__(self):
         return "bvh mocap file with these coils: {}".format(str(self.marker_names))
@@ -417,14 +299,15 @@ class BVHParser(MocapParent):
     def read_header(self):
         """Read the BVH header and put in neutral format."""
 
-        line = self.file.readline()     # First line should be 'HIERARCHY'
+        f = self.file           # keep file open after execution
+        line = f.readline()     # First line should be 'HIERARCHY'
 
         # while in the header
         while not line.startswith("MOTION"):
             strippedline = line.strip(' \t\n\r')
 
             # capture root/joint names sequentially
-            if strippedline.startswith('ROOT') or strippedline.startswith('JOINT'):  # get a marker name
+            if strippedline.startswith('ROOT') or strippedline.startswith('JOINT'): # get a marker name
                 _, name = strippedline.split()
                 self.marker_names.append(name)
 
@@ -437,7 +320,7 @@ class BVHParser(MocapParent):
                 self.min_dimensions = num_channels if num_channels < self.min_dimensions else self.min_dimensions
                 self.marker_channels.append(channelline[2:])  # append a tuple of the channels for the marker
 
-            line = self.file.readline()
+            line = f.readline()
         self.min_channels = len(self.marker_channels)
 
         # act based on the header data:
@@ -459,7 +342,7 @@ class BVHParser(MocapParent):
             self.motion_lines_read += 1
             self.latest_timestamp += self.frame_time
 
-            # update Component6D/Coil objects from file
+            # update Component/Coil objects from file
             measurements = nextline.rstrip('\n')
             measurements = measurements.split('\t')
             #print("measurement line is", measurements)
@@ -472,15 +355,13 @@ class BVHParser(MocapParent):
                 measurements = measurements[len(ch):]
 
             print("about to alter component", str(self.component)[:50])
-            self.component.coils = [CoilBuilder6D.build_from_mapping(self.mappings, coilvals)
-                                for coilvals in measurements_by_coil]
 
-            # for i, coil in enumerate(self.component.coils):
-            #   #  print("giving a coil a measurement,", measurements_by_coil[i])
-            #     coil.set_args_to_vars(measurements_by_coil[i])
-            # # wrap Component6D in Wave protocol to data packet level and return
+            for i, coil in enumerate(self.component.coils):
+              #  print("giving a coil a measurement,", measurements_by_coil[i])
+                coil.set_args_to_vars(measurements_by_coil[i])
+            # wrap Component in Wave protocol to data packet level and return
 
-            return 3, DataFrame(components=[self.component]).pack_all(), None
+            return 3, DataFrame(components=[self.component]).pack_components_to_df(), None
 
         # no more data available (static, give up)
         else:
@@ -496,14 +377,16 @@ class TSVParser(MocapParent):
         """ Initialise the TSV Parser object."""
 
         super().__init__(filename)
+        self.read_header()
+        self.update_xml_initial()
 
         print(self.marker_channels)
 
         # mapping of TSV is sens_id, sens_status, x, y, z, q0, qx, qy, qz
         # desired order is q0, qx, qy, qz, x, y, z
-        self.mappings = Mapping(mapping_as_list=[5, 6, 7, 8, 2, 3, 4])
+        self.mappings = [5, 6, 7, 8, 2, 3, 4]
 
-        self.component = Component6D(fileparser=self)  # builds first component objects within init
+        self.component = Component(fileparser=self)  # builds first component objects within init
 
     def __str__(self):
         return "tsv mocap file"
@@ -565,7 +448,7 @@ class TSVParser(MocapParent):
         if self.motion_lines_read < self.max_num_frames:
             self.motion_lines_read += 1
 
-            # update Component6D/Coil objects from file
+            # update Component/Coil objects from file
             # read the next line, split into meta (eg timestamp etc) and data
             if self.file.closed:
                 print(self.file_name)
@@ -573,10 +456,8 @@ class TSVParser(MocapParent):
                 print('THE FILE IS CLOSED WHY?')
                 raise Exception
             measurements = self.file.readline().rstrip('\n')
-            meta, all_measurements = measurements.split('\t')[:self.extra_left_fields], \
-                                     measurements.split('\t')[self.extra_left_fields:]
-            measurements_by_coil = [all_measurements[x:x+self.fields_per_sensor]
-                                    for x in range(0, len(all_measurements), self.fields_per_sensor)]
+            meta, all_measurements = measurements.split('\t')[:self.extra_left_fields], measurements.split('\t')[self.extra_left_fields:]
+            measurements_by_coil = [all_measurements[x:x+self.fields_per_sensor] for x in range(0, len(all_measurements), self.fields_per_sensor)]
 
             # print('@@@', self.fields_per_sensor,len(measurements_by_coil[0]))
             # get the timestamp, convert to ms if seconds
@@ -584,22 +465,17 @@ class TSVParser(MocapParent):
             timestamp = MocapParent.timestamp_to_microsecs(timestamp)# timestamp now in microseconds
 
             print("about to alter component", str(self.component)[:50])
-            print('measurements_by_coil:', measurements_by_coil)
-            self.component.coils = [CoilBuilder6D.build_from_mapping(self.mappings, coilvals)
-                                    for coilvals in measurements_by_coil]
 
-
-            #old
-            # for coilvals, coilobj in zip(measurements_by_coil, self.component.coils):
-            #     floatvals = MocapParent.make_float_or_zero(coilvals)
-            #     # note that coilvals is  id, status, x, y, z, q0, qx, qy, qz
-            #     coilobj.set_args_to_vars(floatvals, mapping=self.mappings)  # self.mapping sorts these to coil order
+            for coilvals, coilobj in zip(measurements_by_coil, self.component.coils):
+                floatvals = MocapParent.make_float_or_zero(coilvals)
+                # note that coilvals is  id, status, x, y, z, q0, qx, qy, qz
+                coilobj.set_args_to_vars(floatvals, mapping=self.mappings)  # self.mapping sorts these to coil order
 
             self.component.timestamp = timestamp # TODO: Are there any other fields that are missing? Eg should I send WavID?
             self.latest_timestamp = timestamp
             #print(self.component)
 
-            return 3, DataFrame(components=[self.component]).pack_all(), self.latest_timestamp
+            return 3, DataFrame(components=[self.component]).pack_components_to_df(), self.latest_timestamp
 
         # no more data available (static, give up)
         else:
@@ -610,20 +486,23 @@ class TSVParser(MocapParent):
 class POSParser(MocapParent):
     """Parses Carstens motion capture files for AG500 and AG501 machines."""
 
-    file_read_mode = 'rb'
-
     def __init__(self, filename):
         """ Initialise the POS parser object. """
         super().__init__(filename)
         # POS file is a binary, read header then open in bytes
+        self.file.close()
+        self.file = open(filename, 'rb')
+        self.read_header()
+        self.update_xml_initial()
 
         # the mapping between pos data format and WAVE order is known
         # mapping of POS is: x, y, z, phi, theta, rms, extra
         # desired order for WAVE is: q0, qx, qy, qz, x, y, z
         # based on standard Euler rotation with order XYZ, if ph
-        self.mappings = Mapping(mapping_as_list=[None, 3, 4, None, 0, 1, 2], angle_type='ELEVATION')
+        self.mappings = [None, 3, 4, None, 0, 1, 2]
+        # TODO: Check in literature that Carstens' phi and theta represent Euler X and Y rotation
 
-        self.component = Component6D(fileparser=self) # build the first component within the mocap object
+        self.component = Component(fileparser=self) # build the first component within the mocap object
 
     def read_header(self):
         """Read the version, channel info and sampling rate from the ASCII header."""
@@ -668,27 +547,24 @@ class POSParser(MocapParent):
         """
         if self.motion_lines_read < self.max_num_frames:
 
-            # update the timestamp
-            timestamp = self.timestamp_to_microsecs(self.motion_lines_read * self.frame_time)
-            self.component.timestamp, self.latest_timestamp = timestamp, timestamp
-            print('BVH latest timestamp:', self.latest_timestamp)
+            measurements = self.file.read(self.bytes_in_frame)
 
+            # there appears to be no timestamp in the data
+            timestamp = MocapParent.timestamp_to_microsecs(self.motion_lines_read * self.frame_time)
 
-            # 7f is floats for: x, y, z, phi, theta, rms, extra
-            measurements_by_coil = struct.iter_unpack('< 7f', self.file.read(self.bytes_in_frame))
-            # for m in measurements_by_coil:
-            #     print('POS measurement by coil:', m)
-            #
-            # print([m for m in measurements_by_coil])
+            # 7I is integers for: x, y, z, phi, theta, rms, extra
+            measurements_by_coil = struct.iter_unpack('< 7f', measurements[:])
+            #print('&&&',measurements_by_coil)
+            for coilvals, coilobj in zip(measurements_by_coil, self.component.coils):
+                floatvals = MocapParent.make_float_or_zero(coilvals)
+                coilobj.set_args_to_vars(floatvals, mapping=self.mappings)
 
-            self.component.coils = [CoilBuilder6D.build_from_mapping(self.mappings, coilvals)
-                                    for coilvals in measurements_by_coil]
-            # print([(self.mappings, coilvals)for coilvals in measurements_by_coil])
-            assert len(self.component.coils) > 0  # at least one coil must be made
+            self.component.timestamp = timestamp
+            self.latest_timestamp = timestamp
+            print('test timestamp!!', self.latest_timestamp)
 
-            # update the line statistics
             self.motion_lines_read += 1
-            return 3, DataFrame(components=[self.component]).pack_all(), None
+            return 3, DataFrame(components=[self.component]).pack_components_to_df(), None
 
         # no more data available (static, give up)
         else:
@@ -703,37 +579,36 @@ if __name__ == "__main__":
         p = POSParser('./data/Example4.pos')
         for i in range(10):
             mf, *_= p.give_motion_frame()
-            print('Motion frame from POSParser', mf)
-        print('ET parameter string for POS Parser', ET.tostring(p.xml_tree.getroot()))
+            print(mf)
+        print(ET.tostring(p.xml_tree.getroot()))
 
 
     if True:  # test TSV parser
         c = TSVParser('./data/Example3.tsv')
         print('motion frame', c.give_motion_frame())
         mf, msg, *_ = c.give_motion_frame()
-        print('Parameter string for TSV parser', ET.tostring(c.xml_tree.getroot()))
+        print(ET.tostring(c.xml_tree.getroot()))
 
     if True:  # test BVH paser
         a = BVHParser('./data/Example2.bvh')
-        print('marker names', len(a.marker_names), a.marker_names)
-        print('marker channels', len(a.marker_channels), a.marker_channels)
-        print('parameter string', ET.tostring(a.xml_tree.getroot()))
+        print(len(a.marker_names), a.marker_names)
+        print(len(a.marker_channels), a.marker_channels)
+        print(ET.tostring(a.xml_tree.getroot()))
 
-        print('num_frames', a.num_frames)
-        print('frame_time', a.frame_time)
-        print('is BVH file closed',a.file.closed)
-        print('position in BVH file', a.file.tell())
+        print(a.num_frames)
+        print(a.frame_time)
+        print(a.file.closed)
+        print(a.file.tell())
 
         mylabels = ['position', 'Yposition' ,'Zposition' ,'Zrotation', 'Xrotation', 'Yrotation']
-        b = Mapping.order_measurements_as_wave(mylabels, nchannels=7)
-        print(mylabels, 'get reordered as', b)
+        b = MocapParent.order_measurements_as_wave(mylabels, nchannels=7)
         print(a.xml_tree.getroot())
 
 
         a = BVHParser('./data/Example2.bvh')
-        print('marker names', len(a.marker_names), a.marker_names)
-        print('marker channels', len(a.marker_channels), a.marker_channels)
+        print(len(a.marker_names), a.marker_names)
+        print(len(a.marker_channels), a.marker_channels)
         print(a.xml_tree.getroot())
 
-        mappings = [Mapping.order_measurements_as_wave(c,nchannels=7) for c in a.marker_channels]
+        mappings = [MocapParent.order_measurements_as_wave(c,nchannels=7) for c in a.marker_channels]
         print(mappings)
